pyproject.toml
[tool.poetry]
name = "sql-auditor"
version = "0.1.0"
description = "LLM-driven SQL auditor for optimization and analysis"
authors = ["SQL Auditor Team"]
readme = "README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
sqlglot = "^23.0.0"
psycopg2-binary = "^2.9.9"
httpx = "^0.25.2"
tenacity = "^8.2.3"
openai = "^1.3.0"
asyncpg = "^0.29.0"
slowapi = "^0.1.9"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
ruff = "^0.1.6"
black = "^23.11.0"
mypy = "^1.7.0"
jupyter = "^1.0.0"
ipykernel = "^6.26.0"

[tool.poetry.scripts]
audit = "backend.app:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "-v --cov=backend --cov-report=term-missing --cov-report=html --maxfail=1 -q"

[tool.coverage.run]
source = ["backend"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

